refactor: bind 'import' through index file
