+++ src/App.jsx
@@ -1,11 +1,9 @@
 import { useReducer } from 'react';
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
 
-import Layout from 'components/@common/Layout';
-import Header from 'components/@common/Header';
-import MainPage from 'page/MainPage';
-import CardAppPage from 'page/CardAddPage';
-import ConfirmationPage from 'page/ConfirmationPage';
+import { Layout, Header } from 'components';
+import { MainPage, CardAddPage, ConfirmationPage } from 'page';
+
 import reducer from 'reducer/card';
 import { ROUTE, BASENAME } from 'constants';
 import { CardContext } from 'hooks/useCardContext';
@@ -24,7 +22,7 @@ const App = () => {
           <Header />
           <Routes>
             <Route path={ROUTE.MAIN} element={<MainPage />} />
-            <Route path={ROUTE.ADD} element={<CardAppPage />} />
+            <Route path={ROUTE.ADD} element={<CardAddPage />} />
             <Route path={ROUTE.CONFIRM + ':id'} element={<ConfirmationPage />} />
             <Route path={ROUTE.NO_MATCHED} element={<Navigate to={ROUTE.MAIN} />} />
           </Routes>
