+++ src/components/index.js
@@ -1,0 +1,9 @@
+export { default as FormInput } from 'components/@common/FormInput';
+export { default as Button } from 'components/@common/Button';
+export { default as Header } from 'components/@common/Header';
+export { default as Layout } from 'components/@common/Layout';
+export { default as Modal } from 'components/@common/Modal';
+export { default as Tooltip } from 'components/@common/Tooltip';
+
+export { default as CardCompany } from 'components/CardCompany';
+export { default as CardPreview } from 'components/CardPreview';
