+++ src/page/index.js
@@ -1,0 +1,3 @@
+export { default as CardAddPage } from 'page/CardAddPage';
+export { default as ConfirmationPage } from 'page/ConfirmationPage';
+export { default as MainPage } from 'page/MainPage';
